Add '.call' as alias for '.run'
With the ability to use '.call' as an alias for '.run' actionable actions can be
chained together as steps in a 'Dry::Transaction' process.  Dry::Transaction
requires that individual steps respond to '.call' similar to Procs/Lambdas.
This allows for us to treat an actionable object as another function in a
transaction chain.