Migrate Actionable to Jenkins
This commit removes the old Travis link from the url which does not exist anymore.

We also then add the files that are needed to run the test suite and rubocop inside of the jenkins pipeline.

I utilized field_struct repo as an example and leaned on Kyle Archer for explanation of our Jenkins / Docker environment.